requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.15
python-dotenv==1.0.0
PyJWT==2.6.0
requests==2.25.0
urllib3==1.26.5
cryptography==41.0.0
httpx==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
sqlalchemy==1.4.46
